[cmake] Do not treat warnings as errors
clang is stricter than gcc here and triggers errors on conversion
warnings, notably when changing signedness or within initializer lists.
This prevents treating warnings as errors.

Note that it would be better to avoid all those type conversions, but
it's a lot of changes for not much benefit at this point